Add class to wrap call to WARP XHR endpoint to fetch collection day information
[Problem]
The collection day tool needs to know when the collection day is and what types of collections will be performed.

[Solution]
The king county collection day website performs an XHR request to get the collection day information, which is returned
in JSON format. Create a class to wrap this logic and return a simple object to interact with that contains only the
needed data.

[Test]
* unit tests