Add Dockerfile to build AWS Lambda in
[Problem]
AWS Lambda runs on Amazon Linux and hence requires Linux compiled python native libraries. The Mac compiled libraries
were causing the Lambda to fail execution.

[Solution]
Build the project using Docker using the **miniconda3** distribution which is based on Debian. The **README.md** was
updated to provide some light instructions on how to use this and how to configure.

[Test]
Followed instructions, uploaded tarball to AWS Lambda, and ran test event which succeeded in sending me the message.