Integrate with new REST collection calendar API
## Problem
WARP is not returning the correct results. The recycling pick up is a week off.

## Solution
There is a new pick up UI on the Seattle utilities website. When putting in the address information, I found that a
different API is now being used. This change integrates with this new API.

In addition to this, the CLI command has been updated to support printing the result instead of sending the result to
SNS.

## Test
* updated unit tests
* tested locally

```
python -m cdl --aws-profile default --print --address '5317 Renton Ave S'
```

* tested by one off test with Lambda